Rough and ready example program that can play pcap files into libjitsi.
Only supports UDP audio and the payload types are hardcoded at the moment.
Also, doesn't use the timing information from the pcap.

But it works for both g711 and SILK.

PCapDatagramSocket implements a socket interface to a pcap file.
PCapStreamConnector uses this to provide a pcap file back stream connector.
PlayRTP brings it all together.