If an H264 offer is received with a packetization-mode value higher than what we support, answer with our supported packetization-mode.
This is wrong as per RFC3984 (packetization-mode must be symmetric), but
is correct from an interop point of view - packetization-mode levels mean
the offerer supports all modes up to and including the sent level. The
offerer should be sending all modes supported. In our case, the CFS is
stripping out duplicate codec offers, so this fix is a workaround to that
behaviour.