SFR 444757 - One-way media for around 10s after bringing call back from hold
If we're processing a Buffer multiple times (e.g. because we inserted PLC or
found FEC data in the packet) then make sure we update the last seen sequence
number to the one we just processed.

The existing code just incremented the last seen number but this doesn't cope
with large gaps in sequence number (e.g. if the call was put on hold)