SFR 452687 Handle CO_E_NOTINITIALIZED error
Handle CO_E_NOTINITIALIZED when calling
IMediaObject_processOutput.  If this happens, try calling CoInitializeEx
and then try again.

The most likely explanation is the underlying OS thread is different to
the one that we were using when we first called CoInitializeEx (or maybe
we forgot to do that entirely), and so it borks.  It might be neater to
try to be more certain about which thread we're on, but actually Java
doesn't really have access to that.

The other approach would be to check we've made the call in the underlying
native code - but that's tricky because we'd want to access Thread Local
Storage, which isn't exposed by MinGW.

(cherry picked from commit 5910bc365c5077f27c54f750974f77bc249dd822)