SFR 455120 - Accession fails to ACK a 200 OK when coming off hold, causing a dropped call.
The main SIP thread processing the OK was blocked trying to set the traffic type
of the outbound media.
Fix is to modify CallPeerMediaHandler to set the traffic type in a new thread.
Also added logging to various related classes.

This commit contains only logging; see the sister commit to AccessionDesktop for the fix.