One way audio when not behind NAT
Cause: INVITE was created with different stream connector to the media
stack (due to pre create work) thus media was being sent to the wrong
port.

Fix: The fix (in AccessionDesktop) is to share the stream connector with
all the precreated media stacks.  However this means that when closing a
media stack we can no longer close the socket as it may be in use by
another stack.